transfering validate_dst_dir_arg function #561
