+++ im_v2/common/data/extract/download_bulk.py
@@ -36,6 +36,7 @@ import im_v2.binance.data.extract.extractor as imvbdexex
 import im_v2.ccxt.data.extract.extractor as imvcdexex
 import im_v2.common.data.extract.extract_utils as imvcdeexut
 import im_v2.common.data.transform.transform_utils as imvcdttrut
+import im_v2.common.data.qa.validate_input_args as imvcdqviar
 import im_v2.crypto_chassis.data.extract.extractor as imvccdexex
 
 _LOG = logging.getLogger(__name__)
@@ -68,22 +69,6 @@ def _parse() -> argparse.ArgumentParser:
     )
     return parser  # type: ignore[no-any-return]
 
-def validate_dst_dir_arg(args: argparse.Namespace) -> None:
-    """
-    Validate destination directory args.
-    
-    :param args: destination directory args
-    """
-    if args.get("dst_dir"):
-        data_format = args.get("data_format")
-        if data_format != 'csv' and data_format != 'parquet':
-            raise RuntimeError(f"--data_format argument cannot be {data_format} it should be either \
-                    'csv' or 'parquet'")
-        if args.get("db_table"):
-            raise RuntimeError(f"--db_table argument present")
-        if args.get("s3_path"):
-            raise RuntimeError(f"--s3_path argument present")
-
 def _run(args: argparse.Namespace) -> None:
     hdbg.init_logger(verbosity=args.log_level, use_exec_path=True)
     args = vars(args)
@@ -109,7 +94,7 @@ def _run(args: argparse.Namespace) -> None:
         )
     else:
         hdbg.dfatal(f"Vendor {vendor} is not supported.")
-    validate_dst_dir_arg(args)
+    # imvcdqviar.validate_dst_dir_arg(args)
     imvcdeexut.download_historical_data(args, exchange)
